requirements.txt
Django==4.2.4
djangorestframework==3.14.0
rest_framework_nested==0.1.0
